+++ packages/x-design-system/demo/src/app.vue
@@ -2,9 +2,12 @@
   <div class="xds:flex xds:flex-col xds:gap-40">
     <Button />
     <ButtonGroup />
+    <FilterFacet />
     <Icon />
     <Highlight />
     <Input />
+    <InputGroup />
+    <Picture />
     <Badge />
     <FilterFacet />
     <ProgressBar />
@@ -20,8 +23,10 @@ import Button from './components/button.vue'
 import FilterFacet from './components/filter-facet.vue'
 import Highlight from './components/highlight.vue'
 import Icon from './components/icon.vue'
+import InputGroup from './components/input-group/index.vue'
 import Input from './components/input.vue'
 import Integration from './components/integration.vue'
+import Picture from './components/picture.vue'
 import ProgressBar from './components/progress-bar.vue'
 import Typography from './components/typography.vue'
 </script>